Created a parent entity
