+++ homeassistant/components/advantage_air/__init__.py
@@ -8,7 +8,11 @@ from advantage_air import ApiError, advantage_air
 from homeassistant.const import CONF_IP_ADDRESS, CONF_PORT
 from homeassistant.exceptions import ConfigEntryNotReady
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
+from homeassistant.helpers.update_coordinator import (
+    CoordinatorEntity,
+    DataUpdateCoordinator,
+    UpdateFailed,
+)
 
 from .const import ADVANTAGE_AIR_RETRY, DOMAIN
 
@@ -72,3 +76,35 @@ async def async_setup_entry(hass, config_entry):
         )
 
     return True
+
+
+class AdvantageAirEntity(CoordinatorEntity):
+    """Parent class for Advantage Air Entities."""
+
+    def __init__(self, instance, ac_key, zone_key=None):
+        """Initialize common aspects of an Advantage Air sensor."""
+        super().__init__(instance["coordinator"])
+        self.async_change = instance["async_change"]
+        self.ac_key = ac_key
+        self.zone_key = zone_key
+
+    @property
+    def _ac(self):
+        return self.coordinator.data["aircons"][self.ac_key]["info"]
+
+    @property
+    def _zone(self):
+        if self.zone_key:
+            return self.coordinator.data["aircons"][self.ac_key]["zones"][self.zone_key]
+        return None
+
+    @property
+    def device_info(self):
+        """Return parent device information."""
+        return {
+            "identifiers": {(DOMAIN, self.coordinator.data["system"]["rid"])},
+            "name": self.coordinator.data["system"]["name"],
+            "manufacturer": "Advantage Air",
+            "model": self.coordinator.data["system"]["sysType"],
+            "sw_version": self.coordinator.data["system"]["myAppRev"],
+        }
